+++ src/method-descriptor.ts
@@ -1,8 +1,9 @@
-import type { Headers, RequestParams, Params } from './types'
+import type { Headers, RequestParams, ParameterEncoderFn, Params } from './types'
 import type { Middleware } from './middleware'
 
 interface MethodDescriptorParams {
   allowResourceHostOverride?: boolean
+  parameterEncoder?: ParameterEncoderFn
   authAttr?: string
   binary?: boolean
   bodyAttr?: string
@@ -23,6 +24,7 @@ interface MethodDescriptorParams {
  * @typedef MethodDescriptor
  * @param {MethodDescriptorParams} params
  *   @param {boolean} params.allowResourceHostOverride
+ *   @param {Function} params.parameterEncoder
  *   @param {String} params.authAttr - auth attribute name. Default: 'auth'
  *   @param {boolean} params.binary
  *   @param {String} params.bodyAttr - body attribute name. Default: 'body'
@@ -40,6 +42,7 @@ interface MethodDescriptorParams {
  */
 export class MethodDescriptor {
   public readonly allowResourceHostOverride: boolean
+  public readonly parameterEncoder: ParameterEncoderFn
   public readonly authAttr: string
   public readonly binary: boolean
   public readonly bodyAttr: string
@@ -56,6 +59,7 @@ export class MethodDescriptor {
 
   constructor(params: MethodDescriptorParams) {
     this.allowResourceHostOverride = params.allowResourceHostOverride || false
+    this.parameterEncoder = params.parameterEncoder || encodeURIComponent
     this.binary = params.binary || false
     this.headers = params.headers
     this.host = params.host